Update PostContext
This moves a post related function from FeatureContext to PostContext
and cleans up coding standards for the class.
